Separate from PenicillinRequest.kt & PenicillinRequestBuilder.kt
